feat: add ILogger DI implementation
