+++ src/BlazorReports/Services/BrowserServices/IBrowserFactory.cs
@@ -1,0 +1,16 @@
+using BlazorReports.Models;
+
+namespace BlazorReports.Services.BrowserServices;
+
+/// <summary>
+/// Factory for creating browser instances
+/// </summary>
+internal interface IBrowserFactory
+{
+  /// <summary>
+  /// Creates a new browser instance
+  /// </summary>
+  /// <param name="browserOptions"></param>
+  /// <returns></returns>
+  ValueTask<Browser> CreateBrowser(BlazorReportsBrowserOptions browserOptions);
+}
